OptikaWebProject/requirements.txt
absl-py==1.3.0
asgiref==3.5.2
astunparse==1.6.3
azure-core==1.26.0
azure-iot-hub==2.6.0
azure-storage-blob==12.14.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cmake==3.24.1.1
cryptography==38.0.1
Django==4.1.1
dlib==19.24.0
facenet-pytorch==2.5.2
firebase-admin==6.0.0
flatbuffers==22.9.24
gast==0.4.0
google-api-core==2.10.2
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos
grpcio
grpcio-status
h5py==3.7.0
httplib2==0.20.4
idna==3.4
isodate==0.6.1
keras==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-vggface==0.6
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
msgpack==1.0.4
msrest==0.7.1
numpy==1.23.4
oauthlib==3.2.1
opencv-python-headless
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
proto-plus
protobuf
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.2
six==1.16.0
sqlparse==0.4.3
tensorboard
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tf-estimator-nightly==2.8.0.dev2021122109
torch==1.12.1
torchvision==0.13.1
typing_extensions==4.4.0
uamqp==1.6.1
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.14.1
